core/python-bridge/requirements.txt
# Core dependencies for Cyber Crack Pro
requests>=2.28.0
aiohttp>=3.8.0
aiogram==2.23.1
redis>=4.5.0
psycopg2-binary>=2.9.5
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
docker>=6.1.0
numpy>=1.24.0
pandas>=2.1.0
openai>=1.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
aiofiles>=23.0.0
aioredis>=2.0.0
asyncio-throttle>=1.0.0
colorama>=0.4.6
tqdm>=4.66.0
click>=8.1.0
pydantic-settings>=2.0.0

# Android-specific (simplified)
androguard>=3.3.5

# Security analysis (simplified)
capstone>=5.0.0
yara-python>=4.2.0

# Logging and monitoring
structlog>=23.0.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Performance
psutil>=5.9.0

# Additional for orchestrator
asyncio>=3.4.3

# Additional for web framework
uvicorn[standard]>=0.24.0